adapting website to new commands

+++ firmware/arduino/robot.h
@@ -1,6 +1,8 @@
 #ifndef ROBOT_H
 #define ROBOT_H
 
+#include <Arduino.h>
+
 class Configuration;
 
 class Actuator {
@@ -11,7 +13,7 @@ class Actuator {
 		Actuator(int ID, Configuration* config);
 		~Actuator();
 		
-		void setSpeed(int speed);
+		void setSpeed(byte speed);
 };	
 
 class Robot {
@@ -26,7 +28,7 @@ class Robot {
 		~Robot();
 		
 		void setup(int N_actuators);
-		void setActuator(int i, int speed);
+		void setActuator(int i, byte speed);
 };
 
 #endif //ROBOT_H
